Prepared Schema Store for new format
