+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/SchemaStore.java
@@ -34,7 +34,7 @@ import org.neo4j.kernel.impl.store.allocator.ReusableRecordsAllocator;
 import org.neo4j.kernel.impl.store.format.RecordFormats;
 import org.neo4j.kernel.impl.store.id.IdGeneratorFactory;
 import org.neo4j.kernel.impl.store.id.IdType;
-import org.neo4j.kernel.impl.store.record.AbstractSchemaRule;
+import org.neo4j.kernel.impl.store.record.SchemaRuleDeserializer2_0to3_1;
 import org.neo4j.kernel.impl.store.record.DynamicRecord;
 import org.neo4j.kernel.impl.store.record.RecordSerializer;
 import org.neo4j.logging.LogProvider;
@@ -71,7 +71,7 @@ public class SchemaStore extends AbstractDynamicStore implements Iterable<Schema
     public List<DynamicRecord> allocateFrom( SchemaRule rule )
     {
         RecordSerializer serializer = new RecordSerializer();
-        serializer = serializer.append( (AbstractSchemaRule)rule );
+        serializer = serializer.append( rule );
         List<DynamicRecord> records = new ArrayList<>();
         DynamicRecord record = getRecord( rule.getId(), nextRecord(), CHECK );
         ReusableRecordsAllocator recordAllocator = new ReusableRecordsAllocator( this.getRecordSize(), record );
@@ -94,6 +94,6 @@ public class SchemaStore extends AbstractDynamicStore implements Iterable<Schema
             throws MalformedSchemaRuleException
     {
         ByteBuffer scratchBuffer = concatData( records, buffer );
-        return AbstractSchemaRule.deserialize( id, scratchBuffer );
+        return SchemaRuleDeserializer2_0to3_1.deserialize( id, scratchBuffer );
     }
 }
